test: Passing all test to specflow-emulator
