+++ src/__features__/Form.steps.ts
@@ -1,0 +1,3 @@
+import { defineFeature } from "specflow-emulator";
+
+defineFeature("./Form.feature");
